ethosu: Read elementwise scalars by type
Read constant scalar operands using their declared width and
signedness. This preserves 8-bit behavior while supporting 16-bit
and 32-bit scalar operands without truncation or incorrect sign
extension.

Part-of: <https://gitlab.freedesktop.org/mesa/mesa/-/merge_requests/42927>

+++ src/gallium/drivers/ethosu/ethosu_ml.h
@@ -112,7 +112,7 @@ struct ethosu_feature_map {
    struct ethosu_tile_box tiles;
    unsigned zero_point;
    float scale;
-   uint16_t scalar;
+   int32_t scalar;
    bool has_scalar;
    uint8_t region;
 };
